Added function that prints the current game board
So I added a method to the Game class called

     def printBoard(self, player):

That, given the player whose turn is active, will print the game board (each zone with the top and bottom card of each zone, and the middle deck and remaining card numbers), the player's current hand, and the cards remaining in each player's hand. Basically, this will be called each turn.